+++ src/main/java/org/jabref/logic/importer/fetcher/BiodiversityLibrary.java
@@ -99,7 +99,7 @@ public class BiodiversityLibrary implements SearchBasedParserFetcher, Customizab
 
     public BibEntry parseBibJSONtoBibtex(JSONObject item, BibEntry entry) throws IOException, URISyntaxException {
         if (item.has("BHLType")) {
-            if (item.getString("BHLType").equals("Part")) {
+            if ("Part".equals(item.getString("BHLType"))) {
                 URL url = getPartMetadataURL(item.getString("PartID"));
                 JSONObject itemsDetails = getDetails(url);
                 entry.setField(StandardField.LANGUAGE, itemsDetails.optString("Language", ""));
@@ -112,7 +112,7 @@ public class BiodiversityLibrary implements SearchBasedParserFetcher, Customizab
                 entry.setField(StandardField.URL, itemsDetails.optString("PartUrl", ""));
             }
 
-            if (item.getString("BHLType").equals("Item")) {
+            if ("Item".equals(item.getString("BHLType"))) {
                 URL url = getItemMetadataURL(item.getString("ItemID"));
                 JSONObject itemsDetails = getDetails(url);
                 entry.setField(StandardField.EDITOR, itemsDetails.optString("Sponsor", ""));